WIP: rework goblins to operate over `TreeT`s
As opposed to adding `GenT` to a monad transformer stack, we simply
use a `State` transformer and perform operations over `GenT`s (and
thus `TreeT`s) as whole entities.

I don't know how to integratet this approach with Generic yet, so
am punting on that for the time being.

This is a first pass and will be revised. It might not work at all.