Goblins: fix some missing instances
The `tinker` definition for `Goblins g (Maybe a)` highlights the
big issue with these new operations which are defined over `Gen`s.
Since our Monads aren't stacked, we are unable to "reach inside"
the `Gen` and pull out a value to `case` on in the `State` monad
context. We can only "go in". In the case of `Maybe` that is
problematic, because we could have a significant datastructure
inside which requires genome, but we wouldn't know in advance how
much genome was required.

One potential way around this is a function something like
```
tweak :: TinkerM g (Gen a -> Gen a)
```
which could be generated ahead of time. Then when we "go in" to
investigate the `Maybe`, we apply the function in the `Just` case
and do nothing to a `Nothing`.

Again, we have to consume genome without knowing whether it is
actually necessary.

Maybe this is fine. It just results in genome bloat. This comes up
every time we have a sum-type, which will be a lot, so we need some
way to handle it.