Add `conjureOrSave` combinator
This can be added to `tinker` definitions to allow the goblin access
to conjuring at each use of `tinker`. Since we seem to use `tinker` as
the entry point into the goblin code, it makes sense to call `conjure`
from `tinker` (however `conjure` will never call `tinker` so we don't
have concerns about infinite recursion).

This combinator also calls `saveInBagOfTricks` on the result enclosed
in the monadic body passed to it, which means we save tinkered-with
values to the bagOfTricks as well.