Rip out Generics
Generic goblins aren't able to save (to the bagOfTricks) all the
intermediate sub-constructors of a datatype. We want to structurally
recur and save all datatypes nested inside the argument datatype.

This isn't easily possible with Generic goblins, because of issues
marshalling between Generic representations and the concrete "normal"
representation which is needed to pass to
`tinkerRummagedOrConjureOrSave`.

This can maybe be revisited later, but for now is not worth the time.